--- src/components/screens/PreRecebimento/AlterarCronograma/components/TabelaEditarCronograma/styles.scss
@@ -1,29 +1,0 @@
-.grid-cronograma {
-  .grid-cronograma-header {
-    > div {
-      font-weight: bold;
-    }
-
-    .title {
-      padding: 10px 19px;
-      min-height: 90px;
-    }
-
-    .crono-header-green {
-      color: #fff;
-      background-color: #198459;
-    }
-  }
-
-  .grid-cronograma-body {
-    > div {
-      padding: 10px 19px;
-      border: 1px solid rgba(241, 241, 241, 0.932);
-    }
-  }
-  .grid-cronograma-header,
-  .grid-cronograma-body {
-    display: grid;
-    grid-template-columns: 20% 9% 9% 22% 15% 25%;
-  }
-}
