+++ src/components/screens/PreRecebimento/AlterarCronograma/index.jsx
@@ -8,10 +8,8 @@ import {
   getSolicitacaoAlteracaoCronograma
 } from "services/cronograma.service";
 import HTTP_STATUS from "http-status-codes";
-import { Form, Field } from "react-final-form";
-import StatefulMultiSelect from "@khanacademy/react-multi-select";
+import { Form, Field, FormSpy } from "react-final-form";
 import DadosCronograma from "../CronogramaEntrega/components/DadosCronograma";
-import TabelaEditarCronograma from "./components/TabelaEditarCronograma";
 import AnaliseDilogDiretoria from "./components/AnaliseDilogDiretoria";
 import { TextArea } from "components/Shareable/TextArea/TextArea";
 import "./styles.scss";
@@ -24,7 +22,6 @@ import {
   SOLICITACAO_ALTERACAO_CRONOGRAMA
 } from "configs/constants";
 import { useHistory } from "react-router-dom";
-import { OnChange } from "react-final-form-listeners";
 import {
   usuarioEhDilogDiretoria,
   usuarioEhDinutreDiretoria,
@@ -32,27 +29,16 @@ import {
 } from "helpers/utilities";
 import { Radio, Spin } from "antd";
 import { FluxoDeStatusCronograma } from "components/Shareable/FluxoDeStatusCronograma";
-
-const opcoesMotivos = [
-  { value: "ALTERAR_DATA_ENTREGA", label: "Data de Entrega" },
-  { value: "ALTERAR_QTD_ALIMENTO", label: "Quantidade Programada" },
-  { value: "OUTROS", label: "Outros" }
-];
-
-const manterDataEQuantidade = (values, values_) => {
-  return (
-    values.motivos &&
-    values.motivos.includes("OUTROS") &&
-    (values_.includes("ALTERAR_QTD_ALIMENTO") ||
-      values_.includes("ALTERAR_DATA_ENTREGA"))
-  );
-};
+import FormEtapa from "../CadastroCronograma/FormEtapa";
 
 export default ({ analiseSolicitacao }) => {
   const urlParams = new URLSearchParams(window.location.search);
   const uuid = urlParams.get("uuid");
   const [restante, setRestante] = useState(undefined);
+  const [duplicados, setDuplicados] = useState([]);
   const [etapas, setEtapas] = useState([{}]);
+  //const [etapasNovas, setEtapasNovas] = useState([{}]);
+  const [initialValues, setInitialValues] = useState({});
   const [cronograma, setCronograma] = useState(null);
   const [aprovacaoDinutre, setAprovacaoDinutre] = useState(null);
   const [aprovacaoDilog, setAprovacaoDilog] = useState(null);
@@ -60,7 +46,6 @@ export default ({ analiseSolicitacao }) => {
     solicitacaoAlteracaoCronograma,
     setSolicitacaoAlteracaoCronograma
   ] = useState(null);
-  const [podeSubmeter, setpodeSubmeter] = useState(false);
   const [carregando, setCarregando] = useState(false);
   const history = useHistory();
 
@@ -75,44 +60,6 @@ export default ({ analiseSolicitacao }) => {
         solicitacaoAlteracaoCronograma.status
       ));
 
-  const checarQuantidadeInformada = values_ => {
-    return !values_.includes("ALTERAR_QTD_ALIMENTO") || restante === 0;
-  };
-
-  const checarDatasInformadas = (values_, values) => {
-    if (values_.includes("ALTERAR_DATA_ENTREGA")) {
-      let podeSubmeter = etapas.every(
-        etapa =>
-          values[`data_programada_${etapa.uuid}`] !== undefined &&
-          values[`data_programada_${etapa.uuid}`] !== null
-      );
-      return podeSubmeter;
-    }
-
-    return true;
-  };
-
-  const handleMotivosChange = (values_, values, form) => {
-    setpodeSubmeter(false);
-    if (manterDataEQuantidade(values, values_)) {
-      values_ = values_.filter(value_ => value_ !== "OUTROS");
-    }
-    if (values_.length !== 0 && values.justificativa) {
-      setpodeSubmeter(
-        checarQuantidadeInformada(values_) &&
-          checarDatasInformadas(values_, values)
-      );
-    }
-    if (values_.includes("OUTROS")) {
-      if (values_.length !== 0 && values.justificativa) {
-        setpodeSubmeter(true);
-      }
-      form.change("motivos", ["OUTROS"]);
-      return;
-    }
-    form.change("motivos", values_);
-  };
-
   const getDetalhes = async () => {
     setCarregando(true);
     if (analiseSolicitacao) {
@@ -126,9 +73,12 @@ export default ({ analiseSolicitacao }) => {
         );
       }
       setSolicitacaoAlteracaoCronograma(responseSolicitacaoCronograma.data);
+      geraInitialValuesSolicitacao(responseSolicitacaoCronograma.data);
       setCronograma(responseCronograma);
       setEtapas(responseCronograma.etapas);
+      //setEtapasNovas(responseCronograma.etapas);
       setRestante(responseCronograma.qtd_total_programada);
+      setDuplicados([]);
       setCarregando(false);
     } else {
       if (uuid) {
@@ -136,11 +86,48 @@ export default ({ analiseSolicitacao }) => {
         if (responseCronograma.status === HTTP_STATUS.OK) {
           setCronograma(responseCronograma.data);
           setEtapas(responseCronograma.data.etapas);
+          geraInitialValuesEtapa(responseCronograma.data);
           setRestante(responseCronograma.data.qtd_total_programada);
           setCarregando(false);
         }
       }
     }
+  };
+
+  const geraInitialValuesSolicitacao = solicitacao => {
+    let values;
+    values = {
+      motivos: solicitacao ? solicitacao.motivo : undefined,
+      justificativa: solicitacao.justificativa,
+      justificativa_cronograma: buscaLogJustificativaCronograma(
+        solicitacao.logs,
+        "cronograma"
+      ),
+      justificativa_dinutre: buscaLogJustificativaCronograma(
+        solicitacao.logs,
+        "dinutre"
+      )
+    };
+    solicitacao.etapas.forEach(e => {
+      values[`quantidade_total_${e.etapa}`] = e.nova_quantidade;
+      values[`data_programada_${e.etapa}`] = e.nova_data_programada;
+    });
+    setInitialValues(values);
+  };
+
+  const geraInitialValuesEtapa = cronograma => {
+    let values = {};
+    cronograma.etapas.forEach((etapa, index) => {
+      values[`empenho_${index}`] = etapa.numero_empenho;
+      values[`etapa_${index}`] = etapa.etapa;
+      values[`parte_${index}`] = etapa.parte;
+      values[`data_programada_${index}`] = etapa.data_programada;
+      values[`quantidade_${index}`] = etapa.quantidade;
+      values[`total_embalagens_${index}`] = etapa.total_embalagens;
+    });
+    values.quantidade_total = cronograma.qtd_total_programada;
+    values.unidade_medida = cronograma.unidade_medida;
+    setInitialValues(values);
   };
 
   const analisadoPelaDinutre = () => {
@@ -149,35 +136,8 @@ export default ({ analiseSolicitacao }) => {
     );
   };
 
-  const verificarQuantidadesPreenchidas = values => {
-    if (values.motivos.includes("ALTERAR_QTD_ALIMENTO")) {
-      return etapas.every(
-        etapa =>
-          values[`quantidade_total_${etapa.uuid}`] !== undefined &&
-          values[`quantidade_total_${etapa.uuid}`] !== null
-      );
-    }
-    return true;
-  };
-
-  const labelDeMotivos = motivos => {
-    let motivo = " ";
-    if (motivos !== undefined) {
-      if (motivos.includes("ALTERAR_DATA_ENTREGA")) {
-        motivo += " Alterar datas de entrega,";
-      }
-      if (motivos.includes("ALTERAR_QTD_ALIMENTO")) {
-        motivo += " Alterar quantidade programada,";
-      }
-      if (motivos.includes("OUTROS")) {
-        motivo += " Outros ";
-      }
-    }
-    return motivo.slice(0, -1);
-  };
-
   const cadastraAlteracao = async values => {
-    const payload = prepararPayloadCronograma(cronograma, values);
+    const payload = prepararPayloadCronograma(cronograma, values, etapas);
     await cadastraSolicitacaoAlteracaoCronograma(payload)
       .then(() => {
         toastSuccess("Solicitação de alteração salva com sucesso!");
@@ -285,6 +245,36 @@ export default ({ analiseSolicitacao }) => {
     // eslint-disable-next-line
   }, [uuid]);
 
+  const onChangeFormSpy = async changes => {
+    let restante = changes.values.quantidade_total;
+    console.log(changes.values);
+    etapas.forEach((e, index) => {
+      if (changes.values[`quantidade_${index}`])
+        restante = restante - changes.values[`quantidade_${index}`];
+    });
+    setRestante(restante);
+    if (etapas.length < 2) return;
+    const partes_etapas = [];
+    etapas.forEach((_, i) => {
+      partes_etapas.push({
+        parte: changes.values[`parte_${i}`],
+        etapa: changes.values[`etapa_${i}`],
+        index: i
+      });
+    });
+    const duplicados = [];
+    partes_etapas.forEach(pe => {
+      if (
+        partes_etapas.filter(
+          pe_ => pe_.parte === pe.parte && pe_.etapa === pe.etapa
+        ).length > 1
+      ) {
+        duplicados.push(pe.index);
+      }
+    });
+    setDuplicados(duplicados);
+  };
+
   return (
     <Spin tip="Carregando..." spinning={carregando}>
       <div className="card mt-3">
@@ -321,92 +311,22 @@ export default ({ analiseSolicitacao }) => {
               )}
               <Form
                 onSubmit={defineSubmit}
-                initialValues={values => {
-                  if (solicitacaoAlteracaoCronograma) {
-                    let dados_iniciais = {
-                      motivos: solicitacaoAlteracaoCronograma
-                        ? solicitacaoAlteracaoCronograma.motivo
-                        : undefined,
-                      justificativa:
-                        solicitacaoAlteracaoCronograma.justificativa,
-                      justificativa_cronograma: buscaLogJustificativaCronograma(
-                        solicitacaoAlteracaoCronograma.logs,
-                        "cronograma"
-                      ),
-                      justificativa_dinutre: buscaLogJustificativaCronograma(
-                        solicitacaoAlteracaoCronograma.logs,
-                        "dinutre"
-                      )
-                    };
-                    solicitacaoAlteracaoCronograma.etapas.forEach(e => {
-                      dados_iniciais[`quantidade_total_${e.etapa}`] =
-                        e.nova_quantidade;
-                      dados_iniciais[`data_programada_${e.etapa}`] =
-                        e.nova_data_programada;
-                    });
-                    return dados_iniciais;
-                  } else {
-                    return values;
-                  }
-                }}
-                render={({ handleSubmit, form, values }) => (
+                initialValues={initialValues}
+                render={({ handleSubmit, values, errors }) => (
                   <form onSubmit={handleSubmit}>
-                    {!solicitacaoAlteracaoCronograma ? (
-                      <div>
-                        <label className="label font-weight-normal">
-                          <span>* </span>Motivo da Solicitação de Alteração
-                        </label>
-                        <Field
-                          component={StatefulMultiSelect}
-                          name="motivos"
-                          disableSearch={true}
-                          hasSelectAll={false}
-                          options={opcoesMotivos}
-                          selected={values.motivos || []}
-                          onSelectedChanged={values_ =>
-                            handleMotivosChange(values_, values, form)
-                          }
-                          overrideStrings={{
-                            search: "Busca",
-                            selectSomeItems: "Selecione o(s) Motivo(s)",
-                            allItemsAreSelected:
-                              "Todos os itens estão selecionados",
-                            selectAll: "Todos"
-                          }}
-                          required
-                        />
-                      </div>
-                    ) : (
-                      <>
-                        <hr />
-                        <p className="head-green">Solicitação de Alteração</p>
-                        <p>
-                          <span className="green">Motivo: </span>{" "}
-                          {labelDeMotivos(values.motivos)}
-                        </p>
-                      </>
-                    )}
-                    {values.motivos &&
-                    (values.motivos.includes("ALTERAR_DATA_ENTREGA") ||
-                      values.motivos.includes("ALTERAR_QTD_ALIMENTO")) ? (
-                      <div>
-                        <TabelaEditarCronograma
-                          etapas={etapas}
-                          solicitacaoAlteracaoCronograma={
-                            solicitacaoAlteracaoCronograma
-                          }
-                          motivos={values.motivos}
-                          cronograma={cronograma}
-                          values={values}
-                          verificarQuantidadesPreenchidas={
-                            verificarQuantidadesPreenchidas
-                          }
-                          setpodeSubmeter={setpodeSubmeter}
-                          restante={restante}
-                          setRestante={setRestante}
-                        />
-                      </div>
-                    ) : null}
+                    <FormSpy
+                      subscription={{ values: true, active: true, valid: true }}
+                      onChange={changes => onChangeFormSpy(changes)}
+                    />
+                    <FormEtapa
+                      etapas={etapas}
+                      setEtapas={setEtapas}
+                      values={values}
+                      duplicados={duplicados}
+                      restante={restante}
+                      unidadeMedida={values.unidade_medida}
+                      fornecedor={true}
+                    />
                     <div className="mt-4">
                       <label className="label font-weight-normal">
                         <span>* </span>Justificativa
@@ -418,19 +338,6 @@ export default ({ analiseSolicitacao }) => {
                         className="input-busca-produto"
                         disabled={solicitacaoAlteracaoCronograma !== null}
                       />
-                      <OnChange name="justificativa">
-                        {value => {
-                          if (value && values.motivos) {
-                            setpodeSubmeter(
-                              checarQuantidadeInformada(values.motivos) &&
-                                checarDatasInformadas(values.motivos, values) &&
-                                verificarQuantidadesPreenchidas(values)
-                            );
-                          } else {
-                            setpodeSubmeter(false);
-                          }
-                        }}
-                      </OnChange>
                     </div>
                     {(usuarioEhDinutreDiretoria() ||
                       (usuarioEhDilogDiretoria() &&
@@ -506,7 +413,7 @@ export default ({ analiseSolicitacao }) => {
                           solicitacaoAlteracaoCronograma
                         }
                         handleSubmit={handleSubmit}
-                        podeSubmeter={podeSubmeter}
+                        podeSubmeter={Object.keys(errors).length === 0}
                         disabledDinutre={disabledDinutre(values)}
                         disabledDilog={disabledDilog(values)}
                       />
