+++ src/components/screens/PreRecebimento/CadastroCronograma/FormEtapa/styles.scss
@@ -1,0 +1,4 @@
+.texto-alimento-faltante {
+  text-align: right;
+  font-size: 12px;
+}
