+++ src/components/screens/PreRecebimento/CadastroCronograma/styles.scss
@@ -46,11 +46,6 @@
       margin: 16px 0px;
     }
 
-    .texto-alimento-faltante {
-      text-align: right;
-      font-size: 12px;
-    }
-
     .vermelho {
       color: #b40c02;
     }
